Move helpers to DOM and fix sigil_x
